fix(unit): A1 PositionLedger + VirtualFill micro pUSD 根治 (R-4 ABI-safe)
P0-2 单位契约收尾的根因修复。retro 复盘定位：unit debt 的根
不在 RM cap，而在 VirtualFill.fill_size_usdc / 两个 PositionLedger
对「micro vs whole pUSD」的口径漂移 —— matcher 产 whole double，
ledger A 存 whole、ledger B ×1e6 存 micro，paper_loop 再 ×1e6 喂 RM，
任何一处漏乘/多乘都 1e6 量级静默烧钱 (踩铁律#3)。

老周 spec (laozhou-a1-positionledger-micro-migration-v1.md) +
老郭裁定：用裸 int64 micro 统一口径 (非 MicroPUSD 包装，ABI 零变)，
7 钳制收口：

  钳-1  micro_pusd.hpp 加 to_micro_pusd(double)：double→micro 唯一入口
        (禁散落手写 (int64)(x*1e6)，负值 round 方向单点守护)
  钳-2  unit_contract_check.py 加 F4：禁 fill_size_usdc × 1e6
        (matcher + 2 ledger 三文件 grep 守护)
  钳-3  字段注释标 micro 口径
  钳-4  bench_e2e_latency MockPosition::apply(int64) 跟随迁移
  钳-5  ADR c6 续章追加 (unit-contract-spec)
  钳-6  P0-1 exposure 红线 unit gate test 跟随 micro
  钳-7  test_virtual_matcher 加 A1_SubOnePusdFill_NotTruncatedToZero
        (<1pUSD fill 不再被 (int64) 截断为 0)

改动：
  - VirtualFill.fill_size_usdc: double → int64 micro (sizeof==120 不变)
  - matcher 产 fill 走 to_micro_pusd(size * rate)
  - risk::PositionLedger / infra::wal::PositionLedger 统一存 micro
    (删 ×1e6 / 删 (int64) 截断)
  - paper_loop FeedRiskGateway 直喂 micro (删 ×1e6)；pnl_fee /1e6 显示
  - ml/hook partial label /1e6 还原 whole pUSD (unit-contract-ok)

回归：A1 受影响 100/100 绿 (matcher/2 ledger/paper/ml/fill_rate)，
全量 serial 1061/1061 绿。pre-push 7 checker 全 [0]
(abi_lock/core_ssot/adr010/secret/unit_contract/risk_enum/worktree_commit)。

红线: R-4 (schema/单位变更走 schema 红线 + 下游审计) / 铁律#3 (回测实盘
单位一致) / R-8.1#3 (字段单位变更必触发下游消费点 audit)。

ABI ref: docs/ARCHIVE/RESEARCH/laoli-laoSun-handshake-v1.md F-00 L1
(A1 ABI 中性：VirtualFill sizeof==120 不变、PositionLedger key/layout 不变，
 仅 fill_size_usdc 字段语义 double→int64 micro，无 static_assert 数值改动 →
 L1 无需三方签，沿用 P0-2 c2/c2b/c3 同源 ABI-neutral 标记。)

+++ src/stcpp/risk/position_ledger.cpp
@@ -41,7 +41,7 @@ void PositionLedger::apply_fill(std::string const& condition_id, std::string con
     // 仅处理成功成交 (BernoulliMissed / SlippageModelReject 不更新仓位)
     if (fill.reject != execution::MatchReject::Ok)
         return;
-    if (fill.fill_size_usdc == 0.0)
+    if (fill.fill_size_usdc == 0)
         return;
 
     // side 语义: VirtualFill 无 side 字段; 调用方约定:
@@ -51,7 +51,8 @@ void PositionLedger::apply_fill(std::string const& condition_id, std::string con
     // 平仓 (SELL) 调用方传 fill_size_usdc 为正值, delta_usdc 负由 is_close 派送:
     // 此处统一用 +fill_size_usdc; 平仓语义由 REST /drain 端 DRAIN state 保证
     // TODO W9 W5: 当 side 信息透传入 VirtualFill 后更新符号逻辑
-    auto const delta_raw = static_cast<std::int64_t>(fill.fill_size_usdc);
+    // A1: fill_size_usdc 已是 int64 micro pUSD, 直存无 cast (消原 (int64)whole 的 <1pUSD 截断丢仓)。
+    auto const delta_raw = fill.fill_size_usdc;
 
     // R-20: 严格透传 as_of_ts_ns, 禁 now()
     auto const ts = fill.as_of_ts_ns;
